correctly copy all input sections into output sections, enable orphaned section errors
previously, the input section description *(.text) failed to capture
most of the .text.[something] sections that rustc produces. this meant
that the stack, intended to be placed after every section, was actually
placed between .text.boot and the rest of the .text sections which were
orphaned and thus placed in identically-named sections at the end of the
binary. similar occurred with .data, .rodata, etc.

we update the useful input section descriptions to wildcards, and add
explicit output section descriptions for every section in the binary.
this allows us to turn on --orphan-handling=error.