remove .cargo directory, pass kernel target via cargo and linker script via build script
we previously had a dilemma:
- using a cargo workspace in the root of the project meant subjecting
  every crate in the workspace to the kernel's .cargo/config.toml which
  specified the custom target json. with foresight, this seemed
  inappropriate because userspace crates should be configured
  independently.
- using a cargo workspace in kernel/ meant the kernel would require a
  Makefile which invoked a Makefile for a53 which invoked svd2rust - an
  awful lot of nesting.

cargo refuses to look at a .cargo directory inside a workspace member
when invoked in the workspace. it will, however, look at the .cargo
directory of a workspace member if it's invoked inside the member's root
directory. we previously relied on this to isolate the kernel's cargo
config to the kernel itself.

this, of course, broke rust-analyzer.

we now use the kernel's existing build.rs, which existed to tell cargo
to re-run when the linker script changed, to additionally pass the
linker script to the linker. building core and specifying the target are
achieved by passing configuration options directly to cargo.