fix rust-analyzer support in runner and buddy-alloc cli
rust-analyzer has two “subsystems” that are configured and operated
independently: errors and warnings (rust-analyzer.check.*), which
actually runs “cargo check” or compatible, and other lsp features
(rust-analyzer.cargo.*), which are handled by rust-analyzer proper
other than build scripts. many of the “check” settings default to
their “cargo” counterparts.

both were broken for the runner, because it was excluded from our
workspace. trying to check it explicitly with --manifest-path worked
once or twice, but it almost always just reports that the runner is
“Fresh” (--verbose), and i have no idea why. this patch fixes that by
including the runner in the workspace.

both were broken for buddy-alloc’s bin target, because neither of them
were configured to enable the cli feature, making main.rs effectively
empty. this patch fixes that by enabling all features.

errors and warnings were broken too, because our config effectively
tries to cargo check the bin target for aarch64-unknown-none. this is
impossible, since we don’t have std in that target triple. while the
errors suggest -Zbuild-std, this doesn’t really solve the problem,
because you just end up with errors about #![feature(restricted_std)],
which seems to be an internal feature.

    $ cargo check -p buddy-alloc --bins --target aarch64-unknown-none
    error[E0463]: can't find crate for `std`
    error[E0601]: `main` function not found in crate `buddy_alloc`

    $ !! -Zbuild-std
    error[E0463]: can't find crate for `panic_abort`
    error[E0601]: `main` function not found in crate `buddy_alloc`
    error[E0658]: use of unstable library feature 'restricted_std'

errors and warnings were also broken for the runner after including
it in our workspace, for similar reasons. we fix these problems by
giving rust-analyzer a custom check command, which checks each package
with the appropriate config.